+++ apps/web/src/app/(protectedPage)/(mainPage)/_components/navigationCardContent.tsx
@@ -9,9 +9,9 @@ import { usePathname, useRouter } from "next/navigation";
 import { useCallback, useMemo } from "react";
 
 import BrandName from "@/components/common/brandName";
-import { isRateLimitErrorResponse } from "@/lib/errorHandlers";
 import type { NavigationConfig } from "@/lib/types/navigation";
 import { api } from "@/trpc/react";
+import { isRateLimitErrorResponse } from "@convoform/rate-limiter";
 import { NavigationLinks } from "./mainNavigation/mainNavigation";
 
 type Props = {
