+++ packages/api/src/lib/utils.ts
@@ -1,3 +1,11 @@
+import {
+  type LimitType,
+  RATE_LIMIT_ERROR_NAME,
+  checkRateLimitThrowError,
+  isRateLimitError,
+} from "@convoform/rate-limiter";
+import { TRPCError } from "@trpc/server";
+
 export const getLastDaysDate = (lastDaysCount: number) => {
   const lastDaysDate = new Date();
   lastDaysDate.setDate(lastDaysDate.getDate() - lastDaysCount);
@@ -25,29 +33,32 @@ export const getCurrentMonthDaysArray = () => {
   );
 };
 
-export function timeAhead(date: string | number | Date) {
-  const seconds = Math.floor(
-    (new Date(date).getTime() - new Date().getTime()) / 1000,
-  );
-  let interval = seconds / 31536000;
-  if (interval > 1) {
-    return `${Math.floor(interval)} years`;
+export const checkRateLimitThrowTRPCError = async ({
+  identifier,
+  message,
+  rateLimitType,
+}: {
+  /** A unique string value to identify user */
+  identifier: string;
+  /** Custom message to send in response */
+  message?: string;
+  /** Limit type E.g. `core`, `AI` etc */
+  rateLimitType?: LimitType;
+}) => {
+  try {
+    await checkRateLimitThrowError({
+      identifier,
+      message,
+      rateLimitType,
+    });
+  } catch (error) {
+    if (error instanceof Error && isRateLimitError(error)) {
+      throw new TRPCError({
+        code: RATE_LIMIT_ERROR_NAME,
+        message: error.message,
+        cause: error.cause,
+      });
+    }
+    throw error;
   }
-  interval = seconds / 2592000;
-  if (interval > 1) {
-    return `${Math.floor(interval)} months`;
-  }
-  interval = seconds / 86400;
-  if (interval > 1) {
-    return `${Math.floor(interval)} days`;
-  }
-  interval = seconds / 3600;
-  if (interval > 1) {
-    return `${Math.floor(interval)} hours`;
-  }
-  interval = seconds / 60;
-  if (interval > 1) {
-    return `${Math.floor(interval)} minutes`;
-  }
-  return `${Math.floor(seconds)} seconds`;
-}
+};
