+++ packages/api/src/router/form.ts
@@ -16,7 +16,7 @@ import {
   DEFAULT_FORM_DESIGN,
   FORM_SECTIONS_ENUMS_VALUES,
 } from "@convoform/db/src/schema/formDesigns/constants";
-import { checkRateLimitThrowTRPCError } from "../lib/rateLimit";
+import { checkRateLimitThrowTRPCError } from "../lib/utils";
 import { createTRPCRouter, protectedProcedure, publicProcedure } from "../trpc";
 
 export const formRouter = createTRPCRouter({
