+++ packages/api/src/router/workspace.ts
@@ -6,7 +6,7 @@ import {
 } from "@convoform/db/src/schema";
 import { z } from "zod";
 
-import { checkRateLimitThrowTRPCError } from "../lib/rateLimit";
+import { checkRateLimitThrowTRPCError } from "../lib/utils";
 import { createTRPCRouter, protectedProcedure } from "../trpc";
 
 export const workspaceRouter = createTRPCRouter({
